misc changes and 5 new trims

+++ src/main/java/me/cooperzilla/trimssmp/utils/CooldownUtils.java
@@ -6,7 +6,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 
 public class CooldownUtils {
-    public static void setCooldown(Player player, String str, Long cooldown, JavaPlugin pl) {
+    public static void setCooldown(Player player, String str, Integer cooldown, JavaPlugin pl) {
         player.setMetadata(str, new FixedMetadataValue(pl, true));
 
         if (!player.hasMetadata("cooldown_multiplier")) {
@@ -18,7 +18,7 @@ public class CooldownUtils {
             public void run() {
                 player.removeMetadata(str, pl);
             }
-        }.runTaskLater(pl, cooldown * player.getMetadata("cooldown_multiplier").get(0).asLong());
+        }.runTaskLater(pl, (long) cooldown * player.getMetadata("cooldown_multiplier").get(0).asInt());
 
 
     }
